updated for vs2005
git-svn-id: svn://svn.sphinxsearch.com/sphinx/trunk@384 406a0c4d-033a-0410-8de8-e80135713968

+++ src/sphinx.cpp
@@ -390,7 +390,7 @@ public:
 		assert ( m_pData );
 
 		m_tState.m_iAttr = 0;
-		m_tState.m_iNow = time ( NULL );
+		m_tState.m_iNow = (DWORD) time ( NULL );
 	}
 
 	/// dtor
